Overhauled pester testing:
Great progress! We're now getting 3 passed tests out of 6 for NodeScripts, which is much better than 0. The issue now is that the mocks need to be in the LabRunner module scope. The error message shows "Could not find Mock for command Invoke-LabWebRequest in module LabRunner" - this means we need to specify -ModuleName LabRunner when creating the mocks.

Let me update the New-StandardMocks function to support module-scoped mocks:

TestHelpers.ps1+24-0
I need to also update the System mocks section:

Read TestHelpers.ps1, lines 271 to 301

Let me update the System mocks to support ModuleName as well: